Refine styling, adjust flexbox, add button
