+++ src/Components/App/App.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import { Route } from 'react-router-dom';
 import './App.css';
 import Ingredients from '../Ingredients/Ingredients';
