+++ src/Components/Card/Card.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+// import React from 'react';
 import './Card.css';
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
