+++ src/Components/Dropdown/Dropdown.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import './Dropdown.css';
 import PropTypes from 'prop-types';
