+++ src/Components/IngredientDetails/IngredientDetails.css
@@ -1,5 +1,5 @@
 .details-wrapper {
-  border: 2px solid purple;
+  /* border: 2px solid purple; */
   height: 85vh;
   display: flex;
   align-items: center;  
@@ -10,29 +10,43 @@
 
 .details-container {
   background-color: rgb(0, 0, 0, .7);
-  /* height: 50vh; */
-  width: 50%;
+  height: 50vh;
+  width: 55%;
   border-radius: 2%;
-  display: flex;
 }
-/* 
+
+.ingredient-info {
+  display: flex;
+  height: 45vh;
+}
+
 .img-wrapper {
   display: flex;
   justify-content: center;
-  border: 2px solid orange;
-  width: 50%;
-} */
+  align-items: center;
+  /* border: 2px solid orange; */
+  width: 40%;
+  padding: 10px;
+}
 
-/* .details-image {
-  width: 25%;
-} */
+.details-image {
+  width: 75%;
+  height: 30vh;
+  border-radius: 5%;
+}
 
 .details-info {
   text-align: center;
-  border: 2px solid red;
-  width: 100%;
+  /* border: 2px solid red; */
+  width: 60%;
   display: flex;
   flex-direction: column;
   justify-content: space-evenly;
-  /* padding: 10px; */
+  padding: 10px;
+}
+
+.button-container {
+  width: 100%;
+  display: flex;
+  justify-content: center;
 }
